requirements.txt
streamlit
python-dotenv
requests
astrapy
groq